Read Logger File from Ori BasicEx1

Read Logger File from Ori BasicEx1

readLogger_Ori_BasicEx1(filepath, infotype = validInfoTypes(),
  blockbegin = "ORI BasicEx1", warn = TRUE, sep = ";", dec = ",",
  colnameDate = "Datum", colnameTime = "Uhrzeit",
  dateformat = .defaultTimeFormat("v5"),
  timeformat = .defaultTimeFormat("v1"))

Arguments

filepath
full path to file generated by automatic sampler

infotype
one or more of the values returned by validInfoTypes

blockbegin
identification of "block begins"; Default: "ORI BasicEx1"

warn
if TRUE, warnings are generated if a block does not contain "Probe"

sep
column separator

dec
decimal character

colnameDate
name of date column

colnameTime
name of time column

dateformat
date format string

timeformat
time format string

References

http://www.origmbh.de/fileadmin/user_upload/pdf/basic_ex_1_mobil/ORI_Basic_Ex1_mobil_de.pdf

Examples

# set path to example file (contained in this package)
(filepath <- grep("Ori_BasicEx1", exampleLoggerFiles(), value = TRUE)[1])

#> [1] "C:/Users/mrustl.KWB/Documents/RProjects/kwb.logger.git/trunk/inst/extdata/Ori/example_Ori_BasicEx1.csv"


# read the "actions" from the file
readLogger_Ori_BasicEx1(filepath, infotype = "actions")

#>    row          myDateTime        Ereignis Fuellmenge  Flasche Probe  Akku  V9
#> 1    2 2013-05-31 14:47:52       Power Off                                    
#> 2    5 2013-05-31 14:51:43        Power On   SW V1.61 HW V2.00                
#> 3    6 2013-05-31 14:51:53       Power Off                                    
#> 4    9 2013-05-31 14:52:02        Power On   SW V1.61 HW V2.00                
#> 5   13 2013-05-31 14:52:40  Programm Start                            12,7   0
#> 6   23 2013-05-31 14:52:40 Flaschenwechsel        0,0                         
#> 7   24 2013-05-31 14:56:35       Power Off                                    
#> 8   27 2013-05-31 14:56:40        Power On   SW V1.61 HW V2.00                
#> 9   31 2013-05-31 14:56:58  Programm Start                            12,7   0
#> 10  41 2013-05-31 14:56:58 Flaschenwechsel        0,0                         
#> 11  42 2013-05-31 14:58:14       Power Off                                    
#> 12  45 2013-05-31 14:59:42        Power On   SW V1.61 HW V2.00                
#> 13  49 2013-05-31 15:00:03  Programm Start                            13,3   0
#> 14  59 2013-05-31 15:00:03 Flaschenwechsel        0,0                         
#> 15  60 2013-05-31 15:50:59       Bluetooth                            13,2   0
#> 16  61 2013-05-31 15:51:32           Probe     1000,0        1     1  12,4   0
#> 17  62 2013-05-31 16:00:59 Flaschenwechsel     1000,0                         
#> 18  63 2013-05-31 16:06:35           Probe     1000,0        2     1  12,4   0
#> 19  65 2013-05-31 16:10:59   Programm Ende     1000,0                         
#> 20  69 2013-05-31 16:10:59  Programm Start                             0,2   0
#> 21  79 2013-05-31 16:10:59 Flaschenwechsel        0,0                         
#> 22  80 2013-05-31 16:11:36           Probe     1000,0        3     1  12,4   0
#> 23  81 2013-05-31 16:40:59 Flaschenwechsel     1000,0                         
#> 24  82 2013-05-31 16:41:36           Probe     1000,0        4     1  12,4   0
#> 25  83 2013-05-31 17:10:59 Flaschenwechsel     1000,0                         
#> 26  84 2013-05-31 17:11:36           Probe     1000,0        5     1  12,5   0
#> 27  85 2013-05-31 17:40:59 Flaschenwechsel     1000,0                         
#> 28  86 2013-05-31 17:41:36           Probe     1000,0        6     1  12,4   0
#> 29  87 2013-05-31 18:10:59 Flaschenwechsel     1000,0                         
#> 30  88 2013-05-31 18:11:36           Probe     1000,0        7     1  12,4   0
#> 31  90 2013-05-31 18:40:59   Programm Ende     1000,0                         
#> 32  91 2013-05-31 18:40:59       Power Off                                    
#> 33  94 2013-06-01 11:24:37        Power On   SW V1.61 HW V2.00                
#>    V10 V11
#> 1         
#> 2         
#> 3         
#> 4         
#> 5    0    
#> 6         
#> 7         
#> 8         
#> 9    0    
#> 10        
#> 11        
#> 12        
#> 13   0    
#> 14        
#> 15   0    
#> 16   0    
#> 17        
#> 18   0    
#> 19        
#> 20   0    
#> 21        
#> 22   0    
#> 23        
#> 24   0    
#> 25        
#> 26   0    
#> 27        
#> 28   0    
#> 29        
#> 30   0    
#> 31        
#> 32        
#> 33        


# read the sample times from the file
readLogger_Ori_BasicEx1(
  filepath, infotype = "times", blockbegin = "ORI BasicEx1 TU Berlin"
)

#>   row          myDateTime Ereignis Fuellmenge Flasche Probe Akku V9 V10
#> 1  61 2013-05-31 15:51:32    Probe       1000       1     1 12.4  0   0
#> 2  63 2013-05-31 16:06:35    Probe       1000       2     1 12.4  0   0
#> 3  80 2013-05-31 16:11:36    Probe       1000       3     1 12.4  0   0
#> 4  82 2013-05-31 16:41:36    Probe       1000       4     1 12.4  0   0
#> 5  84 2013-05-31 17:11:36    Probe       1000       5     1 12.5  0   0
#> 6  86 2013-05-31 17:41:36    Probe       1000       6     1 12.4  0   0
#> 7  88 2013-05-31 18:11:36    Probe       1000       7     1 12.4  0   0


# read both at the same time
x <- readLogger_Ori_BasicEx1(
  filepath, blockbegin = "ORI BasicEx1 TU Berlin"
)

# examine the list structure of the result
str(x)

#> List of 2
#>  $ actions:'data.frame':	33 obs. of  10 variables:
#>   ..$ row       : int [1:33] 2 5 6 9 13 23 24 27 31 41 ...
#>   ..$ myDateTime: chr [1:33] "2013-05-31 14:47:52" "2013-05-31 14:51:43" "2013-05-31 14:51:53" "2013-05-31 14:52:02" ...
#>   ..$ Ereignis  : chr [1:33] "Power Off" "Power On" "Power Off" "Power On" ...
#>   ..$ Fuellmenge: chr [1:33] "" "SW V1.61" "" "SW V1.61" ...
#>   ..$ Flasche   : chr [1:33] "" "HW V2.00" "" "HW V2.00" ...
#>   ..$ Probe     : chr [1:33] "" "" "" "" ...
#>   ..$ Akku      : chr [1:33] "" "" "" "" ...
#>   ..$ V9        : chr [1:33] "" "" "" "" ...
#>   ..$ V10       : chr [1:33] "" "" "" "" ...
#>   ..$ V11       : chr [1:33] "" "" "" "" ...
#>  $ times  :'data.frame':	7 obs. of  9 variables:
#>   ..$ row       : int [1:7] 61 63 80 82 84 86 88
#>   ..$ myDateTime: chr [1:7] "2013-05-31 15:51:32" "2013-05-31 16:06:35" "2013-05-31 16:11:36" "2013-05-31 16:41:36" ...
#>   ..$ Ereignis  : chr [1:7] "Probe" "Probe" "Probe" "Probe" ...
#>   ..$ Fuellmenge: num [1:7] 1000 1000 1000 1000 1000 1000 1000
#>   ..$ Flasche   : int [1:7] 1 2 3 4 5 6 7
#>   ..$ Probe     : int [1:7] 1 1 1 1 1 1 1
#>   ..$ Akku      : num [1:7] 12.4 12.4 12.4 12.4 12.5 12.4 12.4
#>   ..$ V9        : int [1:7] 0 0 0 0 0 0 0
#>   ..$ V10       : int [1:7] 0 0 0 0 0 0 0