Read Logger File from PCE TDS100

Read logger file of flow meter TDS 100 by PCE Deutschland GmbH

readLogger_PCE_TDS100(txt, timeformat = .defaultTimeFormat("v3"))

Arguments

txt
full path to text file generated by logger

timeformat
Date and time format used in the file. Default: "%y-%m-%d %H:%M:%S"

Value

data frame with columns tstamp, flow, flowunit, vel, velunit, UP, DN, Q. Duplicate lines are removed. The data frame has an attribute metadata containing the meta data of the file, as returned by .getMetadata

References

http://www.warensortiment.de/bedienung/ba_durchflussmessgeraet-pce-tds100h_de-v-1-1.pdf

http://www.industrial-needs.com/manual/manual-pce-tds-100h.pdf

Examples

# set path to example file (contained in this package)
filepaths <- grep("PCE_TDS100", exampleLoggerFiles(), value = TRUE)

# read a file containing metadata
x1 <- readLogger_PCE_TDS100(filepaths[1]) # warnings about duplicate timestamps

#> Warning: There are 62 duplicate timestamps:
#>   2011-04-14 12:56:53
#>   2011-04-14 12:56:58
#>   2011-04-14 12:57:03
#>   2011-04-14 12:57:08
#>   2011-04-14 12:57:13
#>   2011-04-14 12:57:18
#>   2011-04-14 12:57:23
#>   2011-04-14 12:57:28
#>   2011-04-14 12:57:33
#>   2011-04-14 12:57:38
#>   2011-04-14 12:57:43
#>   2011-04-14 12:57:48
#>   2011-04-14 12:57:53
#>   2011-04-14 12:57:58
#>   2011-04-14 12:58:03
#>   2011-04-14 12:58:08
#>   2011-04-14 12:58:13
#>   2011-04-14 12:58:18
#>   2011-04-14 12:58:23
#>   2011-04-14 12:58:28
#>   2011-04-14 12:58:33
#>   2011-04-14 12:58:38
#>   2011-04-14 12:58:43
#>   2011-04-14 12:58:48
#>   2011-04-14 12:58:53
#>   2011-04-14 12:58:58
#>   2011-04-14 12:59:03
#>   2011-04-14 12:59:08
#>   2011-04-14 12:59:13
#>   2011-04-14 12:59:18
#>   2011-04-14 12:59:23
#>   2011-04-14 12:59:28
#>   2011-04-14 12:59:33
#>   2011-04-14 12:59:38
#>   2011-04-14 12:59:43
#>   2011-04-14 12:59:48
#>   2011-04-14 12:59:53
#>   2011-04-14 12:59:58
#>   2011-04-14 13:00:03
#>   2011-04-14 13:00:08
#>   2011-04-14 13:00:13
#>   2011-04-14 13:00:18
#>   2011-04-14 13:00:23
#>   2011-04-14 13:00:28
#>   2011-04-14 13:00:33
#>   2011-04-14 13:00:38
#>   2011-04-14 13:00:43
#>   2011-04-14 13:00:48
#>   2011-04-14 13:00:53
#>   2011-04-14 13:00:58
#>   2011-04-14 13:01:03
#>   2011-04-14 13:01:08
#>   2011-04-14 13:01:13
#>   2011-04-14 13:01:18
#>   2011-04-14 13:01:23
#>   2011-04-14 13:01:28
#>   2011-04-14 13:01:33
#>   2011-04-14 13:01:38
#>   2011-04-14 13:01:43
#>   2011-04-14 13:01:48
#>   2011-04-14 13:01:53
#>   2011-04-14 13:01:58

# read a file not containing metadata
x2 <- readLogger_PCE_TDS100(filepaths[2]) # warning about missing meta data

#> Warning: There are 62 duplicate timestamps:
#>   2011-04-14 12:56:53
#>   2011-04-14 12:56:58
#>   2011-04-14 12:57:03
#>   2011-04-14 12:57:08
#>   2011-04-14 12:57:13
#>   2011-04-14 12:57:18
#>   2011-04-14 12:57:23
#>   2011-04-14 12:57:28
#>   2011-04-14 12:57:33
#>   2011-04-14 12:57:38
#>   2011-04-14 12:57:43
#>   2011-04-14 12:57:48
#>   2011-04-14 12:57:53
#>   2011-04-14 12:57:58
#>   2011-04-14 12:58:03
#>   2011-04-14 12:58:08
#>   2011-04-14 12:58:13
#>   2011-04-14 12:58:18
#>   2011-04-14 12:58:23
#>   2011-04-14 12:58:28
#>   2011-04-14 12:58:33
#>   2011-04-14 12:58:38
#>   2011-04-14 12:58:43
#>   2011-04-14 12:58:48
#>   2011-04-14 12:58:53
#>   2011-04-14 12:58:58
#>   2011-04-14 12:59:03
#>   2011-04-14 12:59:08
#>   2011-04-14 12:59:13
#>   2011-04-14 12:59:18
#>   2011-04-14 12:59:23
#>   2011-04-14 12:59:28
#>   2011-04-14 12:59:33
#>   2011-04-14 12:59:38
#>   2011-04-14 12:59:43
#>   2011-04-14 12:59:48
#>   2011-04-14 12:59:53
#>   2011-04-14 12:59:58
#>   2011-04-14 13:00:03
#>   2011-04-14 13:00:08
#>   2011-04-14 13:00:13
#>   2011-04-14 13:00:18
#>   2011-04-14 13:00:23
#>   2011-04-14 13:00:28
#>   2011-04-14 13:00:33
#>   2011-04-14 13:00:38
#>   2011-04-14 13:00:43
#>   2011-04-14 13:00:48
#>   2011-04-14 13:00:53
#>   2011-04-14 13:00:58
#>   2011-04-14 13:01:03
#>   2011-04-14 13:01:08
#>   2011-04-14 13:01:13
#>   2011-04-14 13:01:18
#>   2011-04-14 13:01:23
#>   2011-04-14 13:01:28
#>   2011-04-14 13:01:33
#>   2011-04-14 13:01:38
#>   2011-04-14 13:01:43
#>   2011-04-14 13:01:48
#>   2011-04-14 13:01:53
#>   2011-04-14 13:01:58

#> Warning: Could not find begin of meta data: Outer Diameter

# examine the structures of the results
str(x1)

#> 'data.frame':	879 obs. of  9 variables:
#>  $ myDateTime: chr  "2011-04-14 12:56:53" "2011-04-14 12:56:58" "2011-04-14 12:57:03" "2011-04-14 12:57:08" ...
#>  $ sys       : chr  "*I" "*I" "*I" "*I" ...
#>  $ flow      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ flowunit  : chr  "m3/h" "m3/h" "m3/h" "m3/h" ...
#>  $ vel       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ velunit   : chr  "m/s" "m/s" "m/s" "m/s" ...
#>  $ UP        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ DN        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ Q         : num  NA NA NA NA NA NA NA NA NA NA ...
#>  - attr(*, "metadata")='data.frame':	1 obs. of  9 variables:
#>   ..$ Outer_Diameter  : chr "   25 mm"
#>   ..$ Wall_Thickness  : chr "   4.65 mm"
#>   ..$ Inner_Diameter  : chr "   15.7 mm"
#>   ..$ Pipe_Material   : chr "  0.Carbon Steel"
#>   ..$ Liner_Material  : chr "  0.No Liner"
#>   ..$ Fluid_Type      : chr "  0.Water"
#>   ..$ Transducer_Type : chr "  Standard-M1"
#>   ..$ Txducer_Mounting: chr "  1. Z-Method"
#>   ..$ Txducers_Spacing: chr "  -3.04965 mm"

str(x2)

#> 'data.frame':	167 obs. of  9 variables:
#>  $ myDateTime: chr  "2011-04-14 12:56:53" "2011-04-14 12:56:58" "2011-04-14 12:57:03" "2011-04-14 12:57:08" ...
#>  $ sys       : chr  "*I" "*I" "*I" "*I" ...
#>  $ flow      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ flowunit  : chr  "m3/h" "m3/h" "m3/h" "m3/h" ...
#>  $ vel       : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ velunit   : chr  "m/s" "m/s" "m/s" "m/s" ...
#>  $ UP        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ DN        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ Q         : num  NA NA NA NA NA NA NA NA NA NA ...

# get meta data from attribute "metadata"
kwb.utils::getAttribute(x1, "metadata")

#>   Outer_Diameter Wall_Thickness Inner_Diameter    Pipe_Material Liner_Material
#> 1          25 mm        4.65 mm        15.7 mm   0.Carbon Steel     0.No Liner
#>   Fluid_Type Transducer_Type Txducer_Mounting Txducers_Spacing
#> 1    0.Water     Standard-M1      1. Z-Method      -3.04965 mm