Read Logger File from LT EDGE M100

Read an input file of water level logger "LT EDGE M100"

read_logger_LT_EDGE_M100(file, timeFormat = .defaultTimeFormat("v4"),
  metaToColumns = FALSE, dbg = TRUE, ...)

Arguments

file
full path to the file (either .xle or .csv)

timeFormat
time format in which the timestamps are given in the file. Default: "yyyy/mm/dd HH:MM:SS". It is assumed that the timestamps represent Berlin normal time = Berlin winter time (no daylight saving)

metaToColumns
if TRUE, metadata are written into the last columns of the returned data frame

dbg
if TRUE (default) debug messages (e.g. "Reading...") are shown

…
arguments passed to read_logger_LT_EDGE_M100_csv or read_logger_LT_EDGE_M100_xle

Value

data frame with columns TimestampInFile (character timestamp as it appeared in the file), DateTimeUTC (POSIXct timestamp in time zone "UTC"), LocalDateTime (POSIXct timestamp in time zone "Europe/Berlin"), UTCOffset (integer Offset in hours between UTC and local time), Level (water level in unit as given in the file header), Temperature (water temperature in unit as given in the file header)