Read File From InSituInc Aquatroll

Read File From InSituInc Aquatroll

readLogger_InSituInc_Aquatroll(csv, headerPattern = NULL,
  timestampFormat = c("%d.%m.%Y %H:%M:%S", "%d.%m.%Y%H:%M:%S"),
  tz = "Etc/GMT+1", maxRowToLookForHeader = 700, model = "",
  dbg = FALSE)

Arguments

csv
full path to logger file

headerPattern
pattern matching the header of the table. See source code for the default.

timestampFormat
vector of possible timestamp formats. See help for the default.

tz
time zone. default: "Etc/GMT+1"

maxRowToLookForHeader
number of first rows in the file to read in advance to look for the header line. Default: 700

model
model of the Aquatroll (either "" or "600")

dbg
if TRUE, debug messages are shown

Value

if model = "600" data frame with columns DateTime, Seconds, Temp, ActCond, SpecCond, TDS, Resis, WaterD, DOC, DOP, Tur, Sal otherwise data frame with columns DateTime, WaterTemp.C, SpecCond.uS

Examples

# set path to example file (contained in this package)
files <- grep("Aquatroll", exampleLoggerFiles(), value=TRUE)

# let's have a look on the file structure (read first 75 lines)
cat(paste(readLines(files[1], 75), collapse = "\n"))

#> Report Date:;08.04.2014 12:24:28
#> Report User Name:;miacsomini
#> Report Computer Name:;KWB45
#> Application:;WinSitu.exe
#> Application Version:;5.6.21.0
#> 
#> 
#> Log File Properties
#> File Name;March19_test_Append_2014-04-08_10-00-41-718.wsl
#> Create Date;08.04.2014 10:00:41
#> 
#> Device Properties
#> Device;Aqua TROLL 100
#> Site;Panke Messstelle Bürgerpark
#> Device Name; 
#> Serial Number;192791
#> Firmware Version;1.10
#> Hardware Version;2
#> Device Address;1
#> Device Comm Cfg;19200,8,Even,1,(Modbus-RTU)
#> Used Memory;26
#> Used Battery;61
#> 
#> Log Configuration
#> ;Log Name;March19_test
#> ;Created By;miacsomini
#> ;Computer Name;KWB45
#> ;Application;WinSitu.exe
#> ;Application Version;5.6.21.0
#> ;Create Date;19.03.2014 16:04:56 Westeuropäische Normalzeit
#> ;Log Setup Time Zone;Westeuropäische Normalzeit
#> ;Notes Size(bytes);4096
#> ;Overwrite when full;Disabled
#> ;Scheduled Start Time;Manual Start
#> ;Scheduled Stop Time;No Stop Time
#> ;Type;Linear
#>  ;Interval;Days: 0 hrs: 00 mins: 15 secs: 00
#> 
#> 
#> 
#> 
#> 
#> Other Log Settings
#> 
#> ;Specific Conductivity Model:;Standard Methods
#> ;;Reference Temperature:;25
#> ;;Temperature Coefficient:;0.0191
#> 
#> 
#> Log Notes:
#> Date and Time;Note
#> 19.03.2014 16:04:59;Sensor SN: 192791  Factory calibration has expired.: 31.07.2012 19:14:39
#> 19.03.2014 16:05:00;Used Battery: 61% Used Memory: 28%   User Name: miacsomini
#> 19.03.2014 16:05:22;Manual Start Command
#> 28.03.2014 10:26:46;Log Download - Used Battery: 61% Used Memory: 28%   User Name: miacsomini
#> 04.04.2014 12:06:06;Log Download - Used Battery: 62% Used Memory: 28%   User Name: miacsomini Interval -  Start Time: 28.03.2014 10:20:22
#> 08.04.2014 10:00:20;Log Download - Used Battery: 62% Used Memory: 28%   User Name: miacsomini Interval -  Start Time: 04.04.2014 12:05:22
#> 
#> 
#> Log Data:
#> Record Count;1892
#> 
#> Sensors;1
#> 
#>  ,1,192791,Conductivity
#> 
#> Time Zone: Westeuropäische Normalzeit/Westeuropäische Sommerzeit
#> 
#> ;;Sensor: Cond                  ;Sensor: Cond                  ;Sensor: Cond                  ;
#> ;Elapsed Time;SN#: 192791                   ;SN#: 192791                   ;SN#: 192791                   ;
#> Date and Time;Seconds     ;Actual Conductivity (µS)      ;Temperature (C)               ;Specific Conductivity (µS)    ;
#> 19.03.2014 16:05:22;       0.000;                         0.000;                        15.388;                         0.000;
#> 19.03.2014 16:20:22;     900.000;                       594.037;                         9.188;                       851.072;
#> 19.03.2014 16:35:22;    1800.000;                       592.425;                         9.034;                       852.341;
#> 19.03.2014 16:50:22;    2700.000;                       591.354;                         9.014;                       851.265;


# now read the file
x <- readLogger_InSituInc_Aquatroll(csv = files[1])

#> Error in .stopIfNotAllColumnsFound(newColumnDescription, headerFields): The following columns could not be found with the given patterns:
#>   SpecCond.uS: 'Specific\s?Conductivity\s?\((<b5>|u)S\)'
#> Available columns:
#>   1. 'Date and Time'
#>   2. 'Seconds     '
#>   3. 'Actual Conductivity (5S)      '
#>   4. 'Temperature (C)               '
#>   5. 'Specific Conductivity (5S)    '
#>   6. 'NA'


# show the first lines
head(x)

#> Error in head(x): object 'x' not found


# Read a file of Aquatroll, model "600"
x600 <- readLogger_InSituInc_Aquatroll(files[2], model = "600")

#> Error in .stopIfNotAllColumnsFound(newColumnDescription, headerFields): The following columns could not be found with the given patterns:
#>   ActCond: 'Actual\s?Conductivity\s?\(<b5>S/cm\)'
#>   SpecCond: 'Specific\s?Conductivity\s?\(<b5>S/cm\)'
#> Available columns:
#>   1. 'Date and Time'
#>   2. 'Seconds     '
#>   3. 'Temperature (C)                         '
#>   4. 'Actual Conductivity (5S/cm)             '
#>   5. 'Specific Conductivity (5S/cm)           '
#>   6. 'Total Dissolved Solids (ppt)            '
#>   7. 'Resistivity (ohm-cm)                    '
#>   8. 'Water Density (g/cm3)                   '
#>   9. 'Dissolved Oxygen (concentration) (mg/L) '
#>   10. 'Dissolved Oxygen (%saturation) (%Sat)   '
#>   11. 'Turbidity (NTU)                         '
#>   12. 'Salinity (PSU)                          '
#>   13. 'NA'


# show the first lines
head(x600)

#> Error in head(x600): object 'x600' not found


# show the structure
str(x600)

#> Error in str(x600): object 'x600' not found